- Change file names - Begin huge cleaning
